+++ singer_sdk/helpers/_flattening.py
@@ -1,4 +1,4 @@
-"""Internal helper library for record flatteting functions."""
+"""Internal helper library for record flattening functions."""
 
 from __future__ import annotations
 
@@ -70,7 +70,7 @@ def flatten_key(key_name: str, parent_keys: list[str], separator: str = "__") ->
             inflection.camelize(inflected_key[reducer_index]),
         )
         inflected_key[reducer_index] = (
-            reduced_key if len(reduced_key) > 1 else inflected_key[reducer_index][0:3]
+            reduced_key if len(reduced_key) > 1 else inflected_key[reducer_index][:3]
         ).lower()
         reducer_index += 1
 
@@ -358,8 +358,8 @@ def _flatten_schema(  # noqa: C901, PLR0912
                 items.append((new_key, next(iter(field_schema.values()))[0]))
 
     # Sort and check for duplicates
-    def _key_func(item):
-        return item[0]  # first item is tuple is the key name.
+    def _key_func(item: tuple[str, dict]) -> str:
+        return item[0]  # first item in tuple is the key name.
 
     sorted_items = sorted(items, key=_key_func)
     for field_name, g in itertools.groupby(sorted_items, key=_key_func):
@@ -451,7 +451,11 @@ def _flatten_record(
     return dict(items)
 
 
-def _should_jsondump_value(key: str, value: t.Any, flattened_schema=None) -> bool:
+def _should_jsondump_value(
+    key: str,
+    value: t.Any,  # noqa: ANN401
+    flattened_schema: dict[str, t.Any] | None = None,
+) -> bool:
     """Return True if json.dump() should be used to serialize the value.
 
     Args:
@@ -465,12 +469,9 @@ def _should_jsondump_value(key: str, value: t.Any, flattened_schema=None) -> boo
     if isinstance(value, (dict, list)):
         return True
 
-    if (
+    return bool(
         flattened_schema
         and key in flattened_schema
         and "type" in flattened_schema[key]
         and set(flattened_schema[key]["type"]) == {"null", "object", "array"}
-    ):
-        return True
-
-    return False
+    )