+++ singer_sdk/sinks/core.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import abc
 import copy
 import datetime
+import importlib.util
 import json
 import time
 import typing as t
@@ -12,8 +13,7 @@ from gzip import GzipFile
 from gzip import open as gzip_open
 from types import MappingProxyType
 
-from dateutil import parser
-from jsonschema import Draft7Validator, FormatChecker
+from jsonschema import Draft7Validator
 
 from singer_sdk.exceptions import MissingKeyPropertiesError
 from singer_sdk.helpers._batch import (
@@ -22,7 +22,12 @@ from singer_sdk.helpers._batch import (
     BatchFileFormat,
     StorageTarget,
 )
-from singer_sdk.helpers._compat import final
+from singer_sdk.helpers._compat import (
+    date_fromisoformat,
+    datetime_fromisoformat,
+    final,
+    time_fromisoformat,
+)
 from singer_sdk.helpers._perftimer import BatchPerfTimer
 from singer_sdk.helpers._typing import (
     DatetimeErrorTreatmentEnum,
@@ -65,7 +70,7 @@ class Sink(metaclass=abc.ABCMeta):
             schema: Schema of the stream to sink.
             key_properties: Primary key of the stream to sink.
         """
-        self.logger = target.logger
+        self.logger = target.logger.getChild(stream_name)
         self.sync_started_at = target.initialized_at
         self._config = dict(target.config)
         self._pending_batch: dict | None = None
@@ -119,7 +124,10 @@ class Sink(metaclass=abc.ABCMeta):
             )
             self._sink_timer.start()
 
-        self._validator = Draft7Validator(schema, format_checker=FormatChecker())
+        self._validator = Draft7Validator(
+            schema,
+            format_checker=Draft7Validator.FORMAT_CHECKER,
+        )
 
     def _get_context(self, record: dict) -> dict:  # noqa: ARG002
         """Return an empty dictionary by default.
@@ -399,7 +407,7 @@ class Sink(metaclass=abc.ABCMeta):
             tz=datetime.timezone.utc,
         ).isoformat()
         record["_sdc_batched_at"] = (
-            context.get("batch_start_time", None)
+            context.get("batch_start_time")
             or datetime.datetime.now(tz=datetime.timezone.utc)
         ).isoformat()
         record["_sdc_deleted_at"] = record.get("_sdc_deleted_at")
@@ -526,8 +534,13 @@ class Sink(metaclass=abc.ABCMeta):
                 date_val = value
                 try:
                     if value is not None:
-                        date_val = parser.parse(date_val)
-                except parser.ParserError as ex:
+                        if datelike_type == "time":
+                            date_val = time_fromisoformat(date_val)
+                        elif datelike_type == "date":
+                            date_val = date_fromisoformat(date_val)
+                        else:
+                            date_val = datetime_fromisoformat(date_val)
+                except ValueError as ex:
                     date_val = handle_invalid_timestamp_in_record(
                         record,
                         [key],
@@ -680,15 +693,24 @@ class Sink(metaclass=abc.ABCMeta):
                     tail,
                     mode="rb",
                 ) as file:
-                    open_file = (
+                    context_file = (
                         gzip_open(file) if encoding.compression == "gzip" else file
                     )
-                    context = {
-                        "records": [
-                            json.loads(line)
-                            for line in open_file  # type: ignore[attr-defined]
-                        ],
-                    }
+                    context = {"records": [json.loads(line) for line in context_file]}  # type: ignore[attr-defined]
+                    self.process_batch(context)
+            elif (
+                importlib.util.find_spec("pyarrow")
+                and encoding.format == BatchFileFormat.PARQUET
+            ):
+                import pyarrow.parquet as pq
+
+                with storage.fs(create=False) as batch_fs, batch_fs.open(
+                    tail,
+                    mode="rb",
+                ) as file:
+                    context_file = file
+                    table = pq.read_table(context_file)
+                    context = {"records": table.to_pylist()}
                     self.process_batch(context)
             else:
                 msg = f"Unsupported batch encoding format: {encoding.format}"