+++ tests/_singerlib/test_messages.py
@@ -8,7 +8,7 @@ import pytest
 from pytz import timezone
 
 import singer_sdk._singerlib as singer
-from singer_sdk._singerlib.messages import format_message
+from singer_sdk.io_base import SingerWriter
 
 UTC = datetime.timezone.utc
 
@@ -19,22 +19,24 @@ def test_exclude_null_dict():
 
 
 def test_format_message():
+    singerwriter = SingerWriter()
     message = singer.RecordMessage(
         stream="test",
         record={"id": 1, "name": "test"},
     )
-    assert format_message(message) == (
+    assert singerwriter.format_message(message) == (
         '{"type": "RECORD", "stream": "test", "record": {"id": 1, "name": "test"}}'
     )
 
 
 def test_write_message():
+    singerwriter = SingerWriter()
     message = singer.RecordMessage(
         stream="test",
         record={"id": 1, "name": "test"},
     )
     with redirect_stdout(io.StringIO()) as out:
-        singer.write_message(message)
+        singerwriter.write_message(message)
 
     assert out.getvalue() == (
         '{"type": "RECORD", "stream": "test", "record": {"id": 1, "name": "test"}}\n'